Switching to chai/mocha for better reporters (and to prepare future mocks of grunt)

+++ test/lib-autoupdate.js
@@ -2,59 +2,55 @@
 
 var grunt = require('grunt');
 var path = require('path');
-var extensionConfigs, dynamicFilename = "grunt-crx-13.3.7.crx";
+var expect = require('chai').expect;
 
 var extensionHelper = require(__dirname + '/../lib/crx.js').init(grunt);
 var autoupdateHelper = require(__dirname + '/../lib/autoupdate.js').init(grunt);
 
-module.exports = {
-  setUp: function(done) {
+describe('lib/autoupdate', function(){
+  var extensionConfigs;
+
+  beforeEach(function(){
     extensionConfigs = {
       "standard": extensionHelper.getTaskConfiguration('test-standard'),
       "codebase": extensionHelper.getTaskConfiguration('test-codebase'),
       "exclude": extensionHelper.getTaskConfiguration('test-exclude')
     };
+  });
 
-    grunt.file.delete("test/data/files/");
-    grunt.file.mkdir("test/data/files/");
+  afterEach(function(){
+    grunt.file.exists("test/data/files/") && grunt.file.delete("test/data/files/");
+  });
 
-    done();
-  },
-  'buildXML': {
-    'without codebase, without update_url': function(test){
-      test.expect(1);
+  it('should not write an autoupdate XML file without codebase and without update_url', function(done){
+    var crx = extensionHelper.createObject(extensionConfigs.standard);
+    crx.manifest.update_url = null;
 
-      var crx = extensionHelper.createObject(extensionConfigs.standard);
-      crx.manifest.update_url = null;
+    autoupdateHelper.buildXML(crx, function(){
+      expect(grunt.file.expand('test/data/files/updates.xml')).to.be.empty;
 
-      autoupdateHelper.buildXML(crx, function(){
-        test.equal(grunt.file.expand('test/data/files/updates.xml').length, 0);
+      done();
+    });
+  });
 
-        test.done();
-      });
-    },
-    'with codebase, without update_url': function(test){
-      test.expect(1);
+  it('should not write an autoupadte XML file with codebase and without update_url', function(done){
+    var crx = extensionHelper.createObject(extensionConfigs.codebase);
+    crx.manifest.update_url = null;
 
-      var crx = extensionHelper.createObject(extensionConfigs.codebase);
-      crx.manifest.update_url = null;
+    autoupdateHelper.buildXML(crx, function(){
+      expect(grunt.file.expand('test/data/files/updates.xml')).to.be.empty;
 
-      autoupdateHelper.buildXML(crx, function(){
-        test.equal(grunt.file.expand('test/data/files/updates.xml').length, 0);
+      done();
+    });
+  });
 
-        test.done();
-      });
-    },
-    'with codebase, with update_url': function(test){
-      test.expect(1);
+  it('should build an autoupdate XML file with codebase and with update_url', function(done){
+    var crx = extensionHelper.createObject(extensionConfigs.codebase);
 
-      var crx = extensionHelper.createObject(extensionConfigs.codebase);
+    autoupdateHelper.buildXML(crx, function(){
+      expect(grunt.file.expand('test/data/files/updates.xml')).to.have.length.of(1);
 
-      autoupdateHelper.buildXML(crx, function(){
-        test.equal(grunt.file.expand('test/data/files/updates.xml').length, 1);
-
-        test.done();
-      });
-    }
-  }
-};
+      done();
+    });
+  });
+});
